completed de State and Unit (with Unit interface)

+++ src/main/java/alchemy/LiquidUnit.java
@@ -4,7 +4,7 @@ import be.kuleuven.cs.som.annotate.*;
 // alles tov spoons want dat hebben we dan nodig om de hoeveelheid van een mixture te bepalen
 
 public enum LiquidUnit implements Unit{
-        DROP(1/8),
+        DROP(1.0/8),
         SPOON(1),
         VIAL(5),
         BOTTLE(15),
@@ -12,14 +12,14 @@ public enum LiquidUnit implements Unit{
         BARREL(1260),
         STOREROOM(6300);
 
-        private final int nominalValue;
+        private final double nominalValue;
 
-        LiquidUnit(int nominalValue) {
+        LiquidUnit(double nominalValue) {
                 this.nominalValue = nominalValue;
         }
 
         @Override
-        public int getNominalValue() {
+        public double getNominalValue() {
                 return nominalValue;
         }
